+++ include/scn/detail/context.h
@@ -40,7 +40,7 @@ namespace scn {
             {
                 return *m_stream;
             }
-            basic_arg<Context> arg(unsigned id) const
+            basic_arg<Context> arg(size_t id) const
             {
                 return m_args.get(id);
             }
@@ -84,7 +84,7 @@ namespace scn {
 
             using arg_type = basic_arg<Context>;
 
-            result<arg_type> do_get_arg(unsigned id)
+            result<arg_type> do_get_arg(size_t id)
             {
                 auto a = m_args.get(id);
                 if (!a && !m_args.get(id - 1)) {
@@ -93,7 +93,7 @@ namespace scn {
                 return a;
             }
 
-            result<arg_type> arg(unsigned id)
+            result<arg_type> arg(size_t id)
             {
                 return m_parse_ctx.check_arg_id(id) ? do_get_arg(id)
                                                     : arg_type();
@@ -136,7 +136,7 @@ namespace scn {
         {
             return this->do_get_arg(this->parse_context().next_arg_id());
         }
-        result<arg_type> arg(unsigned id)
+        result<arg_type> arg(size_t id)
         {
             return this->do_get_arg(id);
         }
